requirements.txt
plotly==4.14.1
pandas==1.1.3
matplotlib==3.3.2
streamlit==0.78.0
seaborn==0.11.0
numpy==1.18.5
Pillow==8.1.2
